bump to PR3466 + add test for noise analysis
