+++ pom.xml
@@ -12,7 +12,7 @@
 		<!--<version>14.0-PR1452</version>-->
 
 		<!-- snapshot == not recommended: rather use PR-labelled release!-->
-		<version>2025.0-PR3444</version>
+		<version>2025.0-PR3466</version>
 	</parent>
